+++ R/targomoOptions.R
@@ -4,8 +4,10 @@
 #' For full details of available options see:
 #' \url{https://docs.targomo.com/core/#/Polygon_Service/post_westcentraleurope_v1_polygon}
 #'
-#' @param travelTimes A list of times, in seconds - each time corresponds to a
-#'     different polygon. Your API key will determine how many you can add.
+#' @param travelTimes A list of times - each time corresponds to a
+#'   different polygon. Your API key will determine how many you can add. Acceptable formats
+#'   are numeric (interpreted as seconds), or a string of the form \code{.h.m.s}. E.g. for
+#'   1 hour, "1h", 90 minutes = "1h30m" or "90m" etc.
 #' @param travelType What mode of transport to use - car, bike, walk or public transport.
 #' @param intersectionMode Whether to calculate the union or intersection of multiple sources.
 #' @param carRushHour Account for rush hour while driving.
@@ -18,7 +20,9 @@
 #'   for transit travel type.
 #' @param transitMaxTransfers,transitEarliestArrival Further transit settings.
 #' @param edgeWeight Should calculations be in "time" or "distance"?
-#' @param maxEdgeWeight The max time or distance to search for routes/times in.
+#' @param maxEdgeWeight The max time or distance to search for routes/times in - acceptable
+#'   formats are numeric (seconds or metres), or a string of the form \code{.h.m.s} for time,
+#'   or \code{.km.m.ml} for distance, where ml represents miles (1609 metres).
 #' @param elevation Account for elevation?
 #' @param serializer Should be "geojson" or "json". See API for details.
 #' @param srid The spatial reference of the returned data.
@@ -60,7 +64,7 @@ targomoOptions = function(travelType = "bike",
   leaflet::filterNULL(
     list(
       travelType = travelType,
-      travelTimes = travelTimes,
+      travelTimes = lapply(travelTimes, formatEdgeWeight, type = edgeWeight),
       intersectionMode = intersectionMode,
       carRushHour = carRushHour,
       walkSpeed = walkSpeed,
@@ -77,7 +81,7 @@ targomoOptions = function(travelType = "bike",
       transitEarliestArrival = transitEarliestArrival,
       transitMaxTransfers = transitMaxTransfers,
       edgeWeight = edgeWeight,
-      maxEdgeWeight = maxEdgeWeight,
+      maxEdgeWeight = formatEdgeWeight(maxEdgeWeight, type = edgeWeight),
       elevation = elevation,
       serializer = serializer,
       srid = srid,