+++ examples/00-test-script.R
@@ -13,6 +13,7 @@ leaflet() %>%
   addTargomoRoutes(source_data = data.frame(lat = 51.5267, lng = -0.1925),
                    target_data = data.frame(lat = 51.542, lng = -0.11),
                    options = targomoOptions(travelType = c("transit", "car", "bike"),
-                                            maxEdgeWeight = 3600),
+                                            maxEdgeWeight = "1h"),
                    drawOptions = routeDrawOptions(bikeDashArray = "5,10",
-                                                  showMarkers = TRUE))
+                                                  showMarkers = TRUE),
+                   verbose = FALSE, progress = TRUE)